// client/src/lib/types/ResourceDto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type ResourceDto = { id: bigint, kind: string, title: string, author: string | null, url: string | null, file_path: string | null, added_at: bigint, last_read_at: bigint | null, status: string, tags: Array<string>, };